refactor: csv analysis in init
