+++ tests/test_analysis/test_analysis_csv_exceptions.py
@@ -6,7 +6,8 @@ import pytest
 from asyncpg import Record
 
 from tests.conftest import RESOURCE_EXCEPTION_ID, RESOURCE_EXCEPTION_TABLE_INDEXES
-from udata_hydra.data_formats.csv_like.analyse import analyse_csv
+from udata_hydra.analysis.helpers import download_from_check
+from udata_hydra.data_formats import Csv
 from udata_hydra.db.resource import Resource
 from udata_hydra.utils.db import get_columns_with_indexes
 
@@ -38,9 +39,10 @@ async def test_exception_analysis(
     resource = await Resource.get(RESOURCE_EXCEPTION_ID)
     assert resource is not None
     assert resource["status"] is None
-
+    
     # Analyse the CSV
-    await analyse_csv(check=check)
+    file = await download_from_check(check, Csv)
+    await file.analyse(check=check)
 
     # Check resource status after analysis
     resource = await Resource.get(RESOURCE_EXCEPTION_ID)
