+++ tests/test_analysis/test_analysis_ogc.py
@@ -13,8 +13,8 @@ from udata_hydra.data_formats import (
     Ogc,
     Wfs,
     Wms,
-    detect_data_format_from_check_or_catalog,
 )
+from udata_hydra.data_formats.detect import detect_data_format_from_check_or_catalog
 
 pytestmark = pytest.mark.asyncio
