+++ udata_hydra/analysis/resource.py
@@ -13,8 +13,8 @@ from udata_hydra.crawl.calculate_next_check import calculate_next_check_date
 from udata_hydra.data_formats import (
     DataFormat,
     Ogc,
-    detect_data_format_from_check_or_catalog,
 )
+from udata_hydra.data_formats.detect import detect_data_format_from_check_or_catalog
 from udata_hydra.db.check import Check
 from udata_hydra.db.resource import Resource
 from udata_hydra.db.resource_exception import ResourceException
