+++ udata_hydra/data_formats/data_format.py
@@ -57,4 +57,4 @@ class DataFormat(ABC):
         return cls.__name__.lower() == format
 
     @abstractmethod
-    def analyse(self, check: dict): ...
+    async def analyse(self, check: dict): ...
